+++ apps/chunithm-scraper/src/constants.ts
@@ -1,0 +1,3 @@
+export const stateStoragePath = "state.json";
+
+export const mobileBaseURL = "https://chunithm-net-eng.com/mobile/";